Move 'rubocop-rspec' dependency into HouseStyle
Rather than include settings in HouseStyle's rspec/.rubocop.yml that
need a third party gem to be added as well, add the `rubocop-rspec`
gem as a dependency and include it ourselves.

We are currently assuming that all users of the `house_style` gem
will be using rspec anyway -- but even if they don't, it's a small
overhead with no further dependencies.