Create a crude Rails generator to start using HouseStyle
For Rails projects, create a generator that creates `.rubocop.yml`
config files in the project root and in the RSpec folder. This assumes
that RSpec is in use.

Usage:

1. Add `gem 'house_style'` to the Gemfile
2. Run `bundle install`
3. Run `rails generate house_style:install`

The generator assumes that neither Rubocop config files exists. If
they do, the user will be prompted to overwrite, ignore or study the
diff, just as for other generators' file conflicts.