Add support for rubocop's forthcoming `TargetRubyVersion`
The next release of rubocop will require an explicit declaration of
which ruby version is being supported within the project's `.rubocop.yml`
file, to ensure that the correct parser etc. is being used.

The variations betwene ruby versions are more about getting the most
accurate parser output, as well as identifying version-specific
enhancements. In general, our rules settings should be agnostic about
the ruby version, so we'll proceed on the basis that it's best for each
project to set its own explicit ruby version.

We can include this prior to the new version's release, as its
inclusion will just be ignored by the current version of rubocop.

To support this, we switch the generator to treat the stored YAML files
as ERb-enhanced templates, then use the RUBY_VERSION of the currently
running instance of Ruby to declare the intended ruby version. Rbenv
or RVM will generally ensure that this constant is the same as the
version cited in `.ruby-version`.

Existing apps using `house_style` can rerun the generators (and merge
any changes if needed) or add the ruby version manually.