Only use major & minor components of RUBY_VERSION
Rubocop's new `TargetRubyVersion` parameter requires only the major and
minor version numbers for the Ruby version it is to check against. So,
for example, if you are running `2.1.3` then `TargetRubyVersion` should
be set to `2.1`. Including the third ('patch'/'tiny') component of the
version number causes an error, so the RUBY_VERSION constant cannot be
used in its raw form.