Only add `db/migrate/.rubocop.yml` if `db/` exists
In a Rails app where the `db/` folder has been deleted (for example,
where the app has no internal persistence and is consuming solely API
data), do not recreate it in order to add a migration-specific rubocop
config file.

Also stop Rubocop reporting a warning when dereferencing the Ruby
version.