Update to Rubocop 0.41.2, remove TargetRubyVersion
Updating to v0.41.2 allows us to remove the code from the Rails
generator that manually calculates the `TargetRubyVersion`. Rubocop now
defaults to whatever version of ruby is specified in `.ruby-version`.

It is possible that some flavours of Ruby (e.g., some JRuby versions)
have incompatible `.ruby-version` files that Rubocop cannot parse. In
those cases, the `TargetRubyVersion` can still be used and will take
precedence -- but that need not be the case for the default usage.