Remove checks on generated and third party files
When code is generated or supplied by third parties, we should not
trigger any rubocop style checks, since we are not in a position to
make alterations to the code without the risk of our changes being
overwritten the next time it is generated. For example, binstubs in
`/bin`, the Ruby-based `db/schema.rb` and any vendorized gems in
`/vendor`.

We exclude these locations from any and all rubocop checks.